fixed mongo field types for weather data

+++ app/module/Perna/src/Perna/Hydrator/Weather/AbstractWeatherDataHydrator.php
@@ -22,8 +22,8 @@ abstract class AbstractWeatherDataHydrator extends AbstractHydrator {
 	public function hydrate( array $data, $object ) {
 		/** @var AbstractWeatherData $object */
 		$object->setWeatherId( (int) $data['weather'][0]['id'] );
-		$object->setWindSpeed( $data['wind']['speed'] ?? $data['speed'] );
-		$object->setCloudiness( $data['clouds']['all'] ?? $data['clouds'] );
+		$object->setWindSpeed( (float) ($data['wind']['speed'] ?? $data['speed']) );
+		$object->setCloudiness( (float) ($data['clouds']['all'] ?? $data['clouds']) );
 		$object->setDateTime( $this->createDate( $data['dt'] ) );
 		return $object;
 	}
